app/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
commonmark==0.9.1
coverage==7.9.2
cryptography==45.0.5
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
Faker==37.6.0
fastapi==0.115.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-Caching==2.3.1
Flask-Limiter==2.8.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
gender-guesser==0.4.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
jellyfish==1.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
networkx==3.3
numpy==2.3.2
ordered-set==4.1.0
orjson==3.11.0
packaging==25.0
pandas==2.2.2
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
rapidfuzz==3.9.6
recordlinkage==0.16
referencing==0.36.2
rich==12.6.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.22
starlette==0.38.6
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
# llama-cpp-python
# huggingface
openai
